[游戏框架] 设计改造 ECS

+++ Unity/Assets/Scripts/Model/Share/Game/Battle/Logic/ILogic.cs
@@ -14,7 +14,7 @@
         void OnDestroy(UnitEntity unitEntity);
 
         // 监听的 ComponentId
-        int[] WatchComponentIds();
+        ushort[] WatchComponentIds();
     }
 
     // 数据处理接口 会修改UnitEntity中的数据
